requirements.txt
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
flask>=3.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
flask-cors>=4.0.0 